Adds Websites
Adds multiple new pages to the website. Also changes have been made to the layout page that keeps the website having a consistent look throughout it. A semi working navigation bar was added. The only two pages that work are Home and About because those are the only two pages that exist thus far.