Added new pages and styles.
New pages were added for the user sign in, and the actually shopping page. The navigation bar was also styled so that it utilizes a drop down menu for the bakery portion of the nav bar. Next images were added on the index page with the intent of creating a slideshow still in the process.