Updated Sign Up
Fixed the problem with creating a customer not working properly. When opened up straight to the browser the post funtion was working properly. However when it was accessed via website control the method would not work. It was because the Home controller was controller a Customer controller function. I switched the methods around to make it work from the proper controller. Now the method works as intended always.