Added new API
A new Web API is created. This one is using identity to eventually use authorization. The API is setup right now to take customers and it works. The login, register and logout are the ones microsoft provided. This update is version control before adding authorization to provide a working version.