Added new Migrations
Took away the migration that Entity Framework provides so I could change the Id to an Int. The basic migration was causing a dropped Identity error when trying to add new migrations. By taking away the provided migration and having my UserContext be the first migration fixed the problem, and now the users database primary key is an integer.