Changes to Navbar
Navbars basic form has been completed. It now displays a logo, and for the mobile version displays a menu bar. Once the menu bar is clicked it'll reveal the menu and a x button. Clicking the x button will close the mobile menu. There is a drawback for the desktop version. If the mobile version closes its menu, then a go to the desktop version, the navigation items will no longer be there. This is an issue with how the classes are used and name.