Added Routing
New pages were added to the react application. This means the application will not be completely SPA. This is due to helping with a conceptual SEO. Routing is done for Homepage, Bakery page and About page. Contact was not done yet due to not knowing if the page is actually going to stick around.